requirements.txt
boto3
bcrypt
PyJWT
flask